fix(minor): rename file useAuthSore to useAuthStore
